Post PR edits:
fixed issues in all the code/yaml files -- turns out the `version:`
tag is for the version of conector to download, not for the version of
the config you are writing.

working on data-pipeline.md tutorial -- document now has content for
the outbound connector section, but it needs a lot more work. Should
be fleshed out tomorrow.

The catfacts-outbout-connector.yaml file is still not quite working
right, but it is close. Just need to get it to push to the SQL
directory correctly.