chore: Hide duplication for connectors docs (#532)
The old connector docs are hidden and have deprecation notices because they have a lot of current inter-site linking.

This also re-adds in the cloud connector page, with an update to the config file in the example. Another pass is needed to verify the tutorials. They should be reviewed to use the newer config and commands as well.

This should be the last of the changes made broken out from previously created PR #507 

Closes #507